# specs/001-network-simulator-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NetLab API
  version: 1.0.0
  description: Shared single-host network-simulator control API.
servers:
  - url: /api/v1
security: []
paths:
...
  /labs/{labId}/exports:
    parameters: [{$ref: '#/components/parameters/LabId'}]
    post:
      operationId: exportLab
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /lab-imports:
    post:
      operationId: importLab
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, description: Redacted LaboratoryExport bundle}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /nodes/{nodeId}:
    parameters: [{$ref: '#/components/parameters/NodeId'}]
    get:
      operationId: getNode
      responses:
        '200': {description: Node, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
    delete:
      operationId: deleteNode
      parameters: [{$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /nodes/{nodeId}/consoles:
    parameters: [{$ref: '#/components/parameters/NodeId'}]
    get:
      operationId: listNodeConsoles
      responses:
        '200': {description: Console descriptors, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ConsoleDescriptor'}}}}}
...
  /tasks/{taskId}:
    get:
      operationId: getTask
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses:
        '200': {description: Task, content: {application/json: {schema: {$ref: '#/components/schemas/OperationTask'}}}}
  /tasks/{taskId}/cancel:
    post:
      operationId: cancelTask
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /labs/{labId}/duplicate:
    post:
      operationId: duplicateLab
      parameters: [{$ref: '#/components/parameters/LabId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name], properties: {name: {type: string, minLength: 1}}}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /tasks:
    get:
      operationId: listTasks
      parameters: [{name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000}}]
      responses:
        '200': {description: Durable operation tasks, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/OperationTask'}}}}}
...
  /nodes/{nodeId}/consoles/{mode}/stream:
    get:
      operationId: streamNodeConsole
      parameters: [{$ref: '#/components/parameters/NodeId'}, {name: mode, in: path, required: true, schema: {enum: [ssh, telnet, vnc]}}, {name: session_id, in: query, required: false, schema: {type: string, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'}}]
      responses:
        '101': {description: WebSocket console stream}
  /nodes/{nodeId}/consoles/{mode}/sessions/{sessionId}:
    delete:
      operationId: closeNodeConsoleSession
      parameters: [{$ref: '#/components/parameters/NodeId'}, {name: mode, in: path, required: true, schema: {enum: [ssh, telnet, vnc]}}, {name: sessionId, in: path, required: true, schema: {type: string, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'}}]
      responses:
        '204': {description: Console session closed or already absent}
...
components:
  parameters:
    LabId: {name: labId, in: path, required: true, schema: {type: string, format: uuid}}
    NodeId: {name: nodeId, in: path, required: true, schema: {type: string, format: uuid}}
    TaskId: {name: taskId, in: path, required: true, schema: {type: string, format: uuid}}
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string}, description: Quoted resource revision}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 8, maxLength: 128}}
  responses:
    AcceptedTask: {description: Accepted, content: {application/json: {schema: {$ref: '#/components/schemas/TaskEnvelope'}}}}
...
  schemas:
...
    Node:
      type: object
      required: [id, laboratory_id, name, kind, revision, desired_state, observed_state]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, name: {type: string}, kind: {enum: [qemu, docker, pc, switch_l2, switch_l3]}, template_version_id: {type: string, format: uuid}, revision: {type: integer}, desired_state: {enum: [stopped, running, deleted]}, observed_state: {enum: [unknown, provisioning, starting, running, stopping, stopped, failed, deleting]}, cpu_count: {type: integer}, cpu_quota_micros: {type: integer}, memory_mib: {type: integer}, storage_gib: {type: integer}, interface_limit: {type: integer}, process_limit: {type: integer}, config: {type: object}, last_error: {$ref: '#/components/schemas/Problem'}}
...
    NodeInterface:
      type: object
      required: [id, node_id, slot, name, driver, mac_address, revision]
      properties: {id: {type: string, format: uuid}, node_id: {type: string, format: uuid}, slot: {type: integer}, name: {type: string}, driver: {type: string}, mac_address: {type: string}, desired_link_id: {type: string, format: uuid}, operational_state: {type: string}, revision: {type: integer}}
    Link:
      type: object
      required: [id, laboratory_id, endpoint_a_id, endpoint_b_id, revision, desired_state, observed_state]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, endpoint_a_id: {type: string, format: uuid}, endpoint_b_id: {type: string, format: uuid}, revision: {type: integer}, desired_state: {type: string}, observed_state: {type: string}}
...
    PortMapping:
      type: object
      required: [id, node_id, protocol, host_address, host_port, guest_address, guest_port, revision, observed_state, created_at]
      properties: {id: {type: string, format: uuid}, node_id: {type: string, format: uuid}, protocol: {enum: [tcp, udp]}, host_address: {type: string}, host_port: {type: integer, minimum: 1, maximum: 65535}, guest_address: {type: string}, guest_port: {type: integer, minimum: 1, maximum: 65535}, revision: {type: integer, minimum: 1}, observed_state: {enum: [pending, active, failed, deleting]}, last_error: {$ref: '#/components/schemas/Problem'}, created_at: {type: string, format: date-time}}
    ConsoleDescriptor:
      type: object
      required: [mode, stream_url, idle_seconds]
      properties: {mode: {enum: [ssh, telnet, vnc]}, stream_url: {type: string}, idle_seconds: {type: integer}, reconnectable: {type: boolean}, subprotocol: {type: string}}
...
    OperationTask:
      type: object
      required: [id, kind, resource_type, resource_id, state, progress_current, progress_total, created_at]
      properties: {id: {type: string, format: uuid}, kind: {type: string}, resource_type: {type: string}, resource_id: {type: string}, requested_revision: {type: integer}, state: {enum: [queued, running, cancelling, succeeded, failed, cancelled]}, progress_current: {type: integer}, progress_total: {type: integer}, result: {type: object}, error: {$ref: '#/components/schemas/Problem'}, created_at: {type: string, format: date-time}, started_at: {type: string, format: date-time}, finished_at: {type: string, format: date-time}}
...
    TaskEnvelope:
      type: object
      required: [task]
      properties: {task: {$ref: '#/components/schemas/OperationTask'}, node: {$ref: '#/components/schemas/Node'}, link: {$ref: '#/components/schemas/Link'}, interface: {$ref: '#/components/schemas/NodeInterface'}, port_mapping: {$ref: '#/components/schemas/PortMapping'}, artifact: {$ref: '#/components/schemas/Artifact'}}
    Problem:
      type: object
      properties: {code: {type: string}, message: {type: string}, retryable: {type: boolean}, resource_type: {type: string}, resource_id: {type: string}, task_id: {type: string, format: uuid}, phase: {type: string}, cleanup: {type: string}, operator_hint: {type: string}, retry_after_seconds: {type: integer, minimum: 1}, details: {type: object}}